pyproject.toml
[tool.poetry]
name = "book-api"
version = "0.1.0"
description = ""
authors = ["0xshodan <hngdean@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-filter = "^23.2"
psycopg = {extras = ["binary"], version = "^3.1.10"}
python-dotenv = "^1.0.0"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
